+++ src/app/(main)/feed/detail/[feedId]/components/FeedDetailMoreMenu.tsx
@@ -1,28 +1,17 @@
 import DeleteCircleIcon from '@public/svg/delete-circle.svg';
 import EditIcon from '@public/svg/edit.svg';
-import { motion } from 'framer-motion';
 import { useParams, useRouter } from 'next/navigation';
 
+import DropDown from '@/components/DropDown';
 import { useOverlay } from '@/hooks';
 
 import FeedDeleteDialog from './FeedDeleteDialog';
 
-type FeedDetailMoreDropdownProps = {
+type FeedDetailMoreMenuProps = {
   open: boolean;
 };
 
-const wrapperVariants = {
-  open: {
-    scaleY: 1,
-    opacity: 1,
-  },
-  closed: {
-    scaleY: 0,
-    opacity: 0,
-  },
-};
-
-export default function FeedDetailMoreMenu({ open }: FeedDetailMoreDropdownProps) {
+export default function FeedDetailMoreMenu({ open }: FeedDetailMoreMenuProps) {
   const router = useRouter();
   const params = useParams();
   const { feedId } = params;
@@ -41,27 +30,20 @@ export default function FeedDetailMoreMenu({ open }: FeedDetailMoreDropdownProps
   ];
 
   return (
-    <motion.div
-      animate={open ? 'open' : 'closed'}
-      initial={wrapperVariants.closed}
-      variants={wrapperVariants}
-      className="absolute right-0 top-full min-w-[148px] origin-top overflow-hidden rounded-xl bg-white shadow-md"
-    >
-      <div className="py-2 text-center text-xs text-gray-500">더보기</div>
-      <motion.ul className="w-full">
-        {listItem.map(({ text, icon, onClick }) => (
-          <motion.li className="border-t border-gray-200">
-            <button
-              type="button"
-              onClick={onClick}
-              className="flex w-full items-center justify-between p-4"
-            >
-              <span className="text-sm">{text}</span>
-              {icon}
-            </button>
-          </motion.li>
-        ))}
-      </motion.ul>
-    </motion.div>
+    <DropDown open={open}>
+      <DropDown.Title>더보기</DropDown.Title>
+      {listItem.map(({ text, icon, onClick }) => (
+        <DropDown.Option>
+          <button
+            type="button"
+            onClick={onClick}
+            className="flex w-full items-center justify-between p-4"
+          >
+            <span className="text-sm">{text}</span>
+            {icon}
+          </button>
+        </DropDown.Option>
+      ))}
+    </DropDown>
   );
 }
